Create bookSlice.js
