--- src/redux/categories/categoriesSlice.js
@@ -1,23 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  books: [],
-};
-
-const bookSlice = createSlice({
-  name: 'books',
-  initialState,
-  reducers: {
-    addBook: (state, action) => {
-      state.books.push(action.payload);
-    },
-    removeBook: (state, action) => {
-      const bookId = action.payload;
-      state.books = state.books.filter((book) => book.id !== bookId);
-    },
-  },
-});
-
-export const { addBook, removeBook } = bookSlice.actions;
-
-export default bookSlice.reducer;
